// enums/chat.enum.ts
export enum ActionKind {
  SET_EMOJI_OPEN = 'SET_EMOJI_OPEN',
  SET_CURRENT_CHAT = 'SET_CURRENT_CHAT',
  SET_CURRENT_MESSAGE = 'SET_CURRENT_MESSAGE',
  SET_MESSAGES = 'SET_MESSAGES',
  SEND_MESSAGE = 'SEND_MESSAGE',
  LOAD_EARLIER_MESSAGES = 'LOAD_EARLIER_MESSAGES',
  LOAD_EARLIER_START = 'LOAD_EARLIER_START',
  SET_IS_TYPING = 'SET_IS_TYPING',
}